Move ml/main.py helper functions to separate file
Add traditional python main function to ml/main.py to reinforce it's use
as an executable. Move ml/main.py helper functions into
ml/amaranth_lib.py to reduce clutter and improve readability and
clarity.
